+++ src/CleanRegex/Internal/Prepared/Parser/Entity/GroupOpen.php
@@ -1,12 +1,19 @@
 <?php
 namespace TRegx\CleanRegex\Internal\Prepared\Parser\Entity;
 
+use TRegx\CleanRegex\Internal\Prepared\Parser\Subpattern;
+
 class GroupOpen implements Entity
 {
-    use TransitiveFlags, PatternEntity;
+    use PatternEntity;
 
     public function pattern(): string
     {
         return '(';
+    }
+
+    public function visit(Subpattern $subpattern): void
+    {
+        $subpattern->pushFlagsIdentity();
     }
 }