+++ ChangeLog.md
@@ -4,7 +4,9 @@ T-Regx Changelog
 Incoming
 --------
 
-* Soon
+* Bug fixes
+    * Corrected a bug in prepared patterns, where closing group after in-pattern (`(?i)`) modifier didn't update the
+      flags properly, resulting in improper parsing of placeholders in `Pattern::inject()`/`Pattern::template()`
 
 Added in 0.20.0
 ---------------